+++ GCTR-3/butterknife/gctr-3-prime-bk.c
@@ -55,7 +55,7 @@ void gctr_3_prime_butterknife_rn(const uint8_t key[GCTR3P_BK_KEY_LEN],
     while (offset < len) {
         size_t remaining = len - offset;
         /* Use only as many branches as needed for the remaining data. */
-        uint8_t b = (uint8_t)((remaining + GCTR3P_BK_N - 1u) / GCTR3P_BK_N);
+        size_t b = (remaining + GCTR3P_BK_N - 1u) / GCTR3P_BK_N;
         if (b > GCTR3P_BK_BRANCHES)
             b = GCTR3P_BK_BRANCHES;
         size_t chunk = (size_t)GCTR3P_BK_N * (size_t)b;
@@ -69,7 +69,7 @@ void gctr_3_prime_butterknife_rn(const uint8_t key[GCTR3P_BK_KEY_LEN],
         gctr3p_bk_set_domain_10(tweakey);
 
         /* X_j = N (fixed for the whole message) */
-        butterknife_256_encrypt(tweakey, stream, N, b);
+        butterknife_256_encrypt(tweakey, stream, N, (uint8_t)b);
 
         for (size_t i = 0; i < take; ++i)
             out[offset + i] = (uint8_t)(in[offset + i] ^ stream[i]);
